+++ crates/wast/src/wast.rs
@@ -1,4 +1,4 @@
-use crate::spectest::link_spectest;
+use crate::spectest::*;
 use anyhow::{anyhow, bail, Context as _, Result};
 use std::fmt::{Display, LowerHex};
 use std::path::Path;
@@ -125,6 +125,8 @@ impl<T> WastContext<T> {
     /// Register "spectest" which is used by the spec testsuite.
     pub fn register_spectest(&mut self) -> Result<()> {
         link_spectest(&mut self.core_linker, &mut self.store)?;
+        #[cfg(feature = "component-model")]
+        link_component_spectest(&mut self.component_linker)?;
         Ok(())
     }